# Add metricing system for benchmarking performance