## Summary
In this PR I am adding a metricing system such that I can begin to properly benchmark the performance of the broker and observe if there are noticeable changes as improvements are made.

Additionally note that I am moving the subscriber and publisher into classes of their own which can then be called programmatically and together with the broker in the main function. This will allow me to provide more standardized measurements where each of the classes are called together instead of manually run tasks in separate windows.

## Why
This will be critical for being able to determine if there are actual any performance benefits to any of the future changes I make.

## How
Added the brokermetrics class which should, going forward, contain all of the functionality related to metricing.

## Trade-offs
I was considering just adding custom metricing to the finbroker class itself however I thought it might be best to separate the metrics into its own class to separate its concerns.

## Testing
```
=== Broker Performance Summary ===
Runtime: 15.008 seconds
Messages received: 164989
Messages sent: 164989
Bytes received: 4289714 bytes
Bytes sent: 4289714 bytes
Throughput: 10993.179 msg/sec

Topic breakdown:
  NYSE_feed: 164989 messages

Processing latency:
  P50: 8875 ns
  P95: 239042 ns
  P99: 472708 ns

=== Processing Latency Over Time ===
Message# | Latency(ns) | Chart
---------|-------------|----------------------------------------
       0 |      84125 | ****************************************
    3299 |      30709 | ******************************
    6598 |       5666 | *****
    9897 |       2958 | **
   13196 |      37959 | *************************************
   16495 |       3041 | ***
   19794 |     147875 | ****************************************
   23093 |      68416 | ****************************************
   26392 |       3875 | ***
   29691 |      87042 | ****************************************
   32990 |       8417 | ********
   36289 |      83917 | ****************************************
   39588 |       2125 | **
   42887 |     128000 | ****************************************
   46186 |       2292 | **
   49485 |       2917 | **
   52784 |     360000 | ****************************************
   56083 |      12084 | ************
   59382 |      74792 | ****************************************
   62681 |       2875 | **
   65980 |       7458 | *******
   69279 |     299834 | ****************************************
   72578 |       2000 | **
   75877 |       3250 | ***
   79176 |     334250 | ****************************************
   82475 |     190500 | ****************************************
   85774 |     122500 | ****************************************
   89073 |       3750 | ***
   92372 |      10583 | **********
   95671 |       2833 | **
   98970 |     125291 | ****************************************
  102269 |     152834 | ****************************************
  105568 |      77166 | ****************************************
  108867 |       6375 | ******
  112166 |     140500 | ****************************************
  115465 |     104833 | ****************************************
  118764 |      10667 | **********
  122063 |      78250 | ****************************************
  125362 |      81417 | ****************************************
  128661 |      72750 | ****************************************
  131960 |       1916 | *
  135259 |       2125 | **
  138558 |      26875 | **************************
  141857 |      27250 | ***************************
  145156 |      52916 | ****************************************
  148455 |     130458 | ****************************************
  151754 |      10875 | **********
  155053 |      43834 | ****************************************
  158352 |       2833 | **
  161651 |      15000 | ***************
  164950 |       2875 | **
  ```

## What's next?
I'll actually need to spend more time on this metricing system to standardize what I'm recording (the number of messages I send through, the number of publishers, the number of subscriber, etc.) and make it more modular, that way future tests will be normalized against a shared test. For the meantime will work on a small feature enhancement to reduce friction and finalize the metrics soon after.